pyproject.toml
[project]
name = "fandango_classroom_test"
version = "0.1.0"
description = "Some test for GitHub classroom and Fandango"
readme = "README.md"
authors = [{ name = "Valentin Huber", email = "git@valentinhuber.me" }]
dependencies = ["fandango-fuzzer>=1.0.6", "pytest>=8.4.2", "pytest-ruff>=0.5"]

[project.scripts]
fandango_classroom_test = "fandango_classroom_test:main"

[tool.pytest.ini_options]
addopts = "--ruff --ruff-format"
